pyproject.toml
[project]
name = "pdf-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "babeldoc>=0.5.16",
    "fastapi>=0.119.1",
    "fastapi-cors>=0.0.6",
    "onnxruntime>=1.23.2",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rapidocr>=3.4.2",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
    "sqlmodel>=0.0.22",
]
